Port graphics renderer from canvas to SVG
Because it would be easier to implement mouse event handlers.
But it costs rendering speed, so we keep the SVG nodes as long as possible.
Specifically, they won't be deleted when the time axis is zoomed or shifted.
However, they will be recreated when the tag filter is modified, since it's
memory-inefficient to keep track of hidden objects.