Prevent tool tip window from being too wide to read
On desktop environment, it's desirable to show many lines of text in a
column with limited width, rather than entire screen width, because the
reader's eye have to travel all the way across the screen to continue
reading to the next line.  Unconstrained "div" element can expand all
the way to the browser window width, so we limit it to 600px.

Also, the tool tip window will try not to overflow the browser window in
the X axis.  If we allow it to overflow, the user have to scroll the
browser horizontally to see all the contents, which is annoying.

I'm not so sure about mobiles.